+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+0.0.4 (????-??-??)
+------------------
+
+- the `set-variable` filter now offers the additional variables `{CURRENT_NAMEEXT}`, `{CURRENT_NAMENOEXT}`,
+  and `{CURRENT_EXT}` if the current item passing through implements `kasperl.api.NameSupporter`
+
+
 0.0.3 (2026-05-05)
 ------------------